+++ src/main/java/net/frozenblock/trailiertales/datagen/loot/TTBlockLootProvider.java
@@ -261,6 +261,9 @@ public final class TTBlockLootProvider extends FabricBlockLootTableProvider {
 		this.dropSelf(TTBlocks.MOSSY_BRICK_WALL);
 
 		this.dropSelf(TTBlocks.POLISHED_RESIN_BLOCK);
+		this.dropSelf(TTBlocks.POLISHED_RESIN_STAIRS);
+		this.add(TTBlocks.POLISHED_RESIN_SLAB, this::createSlabItemTable);
+		this.dropSelf(TTBlocks.POLISHED_RESIN_WALL);
 		this.dropSelf(TTBlocks.CRACKED_RESIN_BRICKS);
 		this.dropSelf(TTBlocks.PALE_MOSSY_RESIN_BRICKS);
 		this.dropSelf(TTBlocks.PALE_MOSSY_RESIN_BRICK_STAIRS);