+++ src/main/java/net/frozenblock/trailiertales/datagen/tag/TTBlockTagProvider.java
@@ -233,6 +233,7 @@ public final class TTBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(TTBlocks.MOSSY_COBBLED_DEEPSLATE_STAIRS)
 			.add(TTBlocks.MOSSY_DEEPSLATE_BRICK_STAIRS)
 			.add(TTBlocks.MOSSY_DEEPSLATE_TILE_STAIRS)
+			.add(TTBlocks.POLISHED_RESIN_STAIRS)
 			.add(TTBlocks.CUT_SANDSTONE_STAIRS)
 			.add(TTBlocks.CUT_RED_SANDSTONE_STAIRS)
 			.add(TTBlocks.CHORAL_END_STONE_STAIRS)
@@ -256,6 +257,7 @@ public final class TTBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(TTBlocks.MOSSY_COBBLED_DEEPSLATE_SLAB)
 			.add(TTBlocks.MOSSY_DEEPSLATE_BRICK_SLAB)
 			.add(TTBlocks.MOSSY_DEEPSLATE_TILE_SLAB)
+			.add(TTBlocks.POLISHED_RESIN_SLAB)
 			.add(TTBlocks.CHORAL_END_STONE_SLAB)
 			.add(TTBlocks.CHORAL_END_STONE_BRICK_SLAB)
 			.add(TTBlocks.END_STONE_SLAB);
@@ -281,6 +283,7 @@ public final class TTBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(TTBlocks.MOSSY_COBBLED_DEEPSLATE_WALL)
 			.add(TTBlocks.MOSSY_DEEPSLATE_BRICK_WALL)
 			.add(TTBlocks.MOSSY_DEEPSLATE_TILE_WALL)
+			.add(TTBlocks.POLISHED_RESIN_WALL)
 			.add(TTBlocks.SMOOTH_SANDSTONE_WALL)
 			.add(TTBlocks.CUT_SANDSTONE_WALL)
 			.add(TTBlocks.SMOOTH_RED_SANDSTONE_WALL)
@@ -516,7 +519,7 @@ public final class TTBlockTagProvider extends FabricTagProvider.BlockTagProvider
 			.add(Blocks.POLISHED_BASALT);
 
 		this.getOrCreateTagBuilder(TTBlockTags.SOUND_POLISHED_RESIN)
-			.add(TTBlocks.POLISHED_RESIN_BLOCK);
+			.add(TTBlocks.POLISHED_RESIN_BLOCK, TTBlocks.POLISHED_RESIN_STAIRS, TTBlocks.POLISHED_RESIN_SLAB, TTBlocks.POLISHED_RESIN_WALL);
 
 		// WILDER WILD